+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v2.1.0
+## 05/11/2020
+
+1. [](#new)
+    * Added Patreon Provider
+    * Added Azure Provider [#1](https://github.com/trilbymedia/grav-plugin-login-oauth2-extras/pull/1)
+
 # v2.0.1
 ## 04/27/2019